Standardize on decimal operators

+++ py_portfolio_index/common.py
@@ -1,6 +1,9 @@
-def print_per(input: float):
+from decimal import Decimal
+
+
+def print_per(input: Decimal):
     return f"{round(input*100,4)}%"
 
 
-def print_money(input: float):
+def print_money(input: Decimal):
     return f"${round(input,2)}"
